% Generated by roxygen2 (4.0.1): do not edit by hand
\name{html_tufte_handout}
\alias{html_tufte_handout}
\title{html format for tufte-handout}
\usage{
html_tufte_handout(self_contained = TRUE, theme = "default",
  lib_dir = NULL, mathjax = "default", pandoc_args = NULL,
  dependency_resolver = html_dependency_resolver, copy_resources = FALSE,
  extra_dependencies = NULL, bootstrap_compatible = FALSE, ...)
}
\description{
html format for tufte-handout
}

