% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pdf_tufte_handout}
\alias{pdf_tufte_handout}
\title{pdf format for tufte-handout}
\usage{
pdf_tufte_handout(fig_width = 6.5, fig_height = 4.5, fig_crop = TRUE,
  fig_caption = TRUE, highlight = "default", keep_tex = FALSE,
  latex_engine = "pdflatex", includes = NULL, pandoc_args = NULL)
}
\description{
pdf format for tufte-handout
}

