% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/html_tufte_handout.R
\name{html_tufte_handout}
\alias{html_tufte_handout}
\title{html format for tufte-handout}
\usage{
html_tufte_handout(self_contained = TRUE, theme = "default",
  lib_dir = NULL, keep_md = FALSE, mathjax = "default",
  pandoc_args = NULL, ...)
}
\arguments{
\item{self_contained}{Include all dependencies}

\item{theme}{Bootstrap theme}

\item{lib_dir}{Local directory to copy assets}

\item{keep_md}{Keep knitr-generated markdown}

\item{mathjax}{Include mathjax, "local" or "default"}

\item{pandoc_args}{Other arguments to pass to pandoc}

\item{...}{Additional function arguments to pass to the base R Markdown HTML
output formatter}
}
\description{
Produces a custom output format function in the style of Edward Tufte's handouts. This
essentially recreates the tufte-handout latex document class using bootstrap.
Main features are plot hooks that put figures in the margin (\code{marginfigure = TRUE}),
creates full-width figures (\code{fig.star = TRUE}), and allows "sidenotes". To create a sidenote,
some raw html is required. Usage is \code{<aside> Content </aside>}. See the
package vignette for more details.
}

