% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pdf_tufte_handout}
\alias{pdf_tufte_handout}
\title{pdf format for tufte-handout}
\usage{
pdf_tufte_handout(fig_width = 6.5, fig_height = 4.5, fig_crop = TRUE,
  highlight = "default", keep_tex = FALSE, latex_engine = "pdflatex",
  includes = NULL, pandoc_args = NULL)
}
\arguments{
\item{fig_width}{Default figure width}

\item{fig_height}{Default figure height}

\item{fig_crop}{Crop figures}

\item{highlight}{Code highlighting}

\item{keep_tex}{Preserve pandoc-generated tex document?}

\item{latex_engine}{Which latex engine to use}

\item{includes}{Named list of additional content to include in the document}

\item{pandoc_args}{Other arguments to pass to pandoc}
}
\description{
Produces a custom output format function in the style of Edward Tufte's handouts. This uses the
tufte-handout latex document class.  Main features are plot hooks that put figures in the margin
(\code{marginfigure = TRUE}),
creates full-width figures (\code{fig.star = TRUE}), and allows "sidenotes". Sidenotes are used
instead of footnotes, for example \code{^[Content]}. See the
package vignette for more details.
}

